cleanup and simplify
