+++ src/agentevals/run/result_builder.py
@@ -13,8 +13,8 @@ from typing import Any, Literal
 from uuid import UUID
 
 from ..config import EvalParams
-from ..runner import RunResult
-from ..storage.models import Result
+from ..runner import MetricResult, RunResult
+from ..storage.models import Result, ResultStatus, compute_result_id
 
 EvaluatorType = Literal["builtin", "code", "remote", "openai_eval"]
 
@@ -29,6 +29,50 @@ def classify_evaluator(metric_name: str, params: EvalParams) -> EvaluatorType:
     return "builtin"
 
 
+def result_from_metric_result(
+    *,
+    run_id: UUID,
+    eval_set_item_id: str,
+    eval_set_item_name: str,
+    trace_id: str | None,
+    evaluator_type: EvaluatorType,
+    metric_result: MetricResult,
+) -> Result:
+    """Project an ADK :class:`MetricResult` onto a persistable :class:`Result`.
+
+    The status mapping treats a non-empty ``error`` field as ``ERRORED`` even
+    when ``eval_status`` would have been ``PASSED`` / ``FAILED``, so
+    downstream consumers can filter on status alone without special-casing
+    the error column.
+    """
+    if metric_result.error:
+        status = ResultStatus.ERRORED
+    else:
+        raw = (metric_result.eval_status or "NOT_EVALUATED").upper()
+        status = {
+            "PASSED": ResultStatus.PASSED,
+            "FAILED": ResultStatus.FAILED,
+        }.get(raw, ResultStatus.SKIPPED)
+
+    latency_ms = int(metric_result.duration_ms) if metric_result.duration_ms is not None else None
+
+    return Result(
+        result_id=compute_result_id(run_id, eval_set_item_id, metric_result.metric_name),
+        run_id=run_id,
+        eval_set_item_id=eval_set_item_id,
+        eval_set_item_name=eval_set_item_name,
+        evaluator_name=metric_result.metric_name,
+        evaluator_type=evaluator_type,
+        status=status,
+        score=metric_result.score,
+        per_invocation_scores=list(metric_result.per_invocation_scores or []),
+        trace_id=trace_id,
+        details=metric_result.details or {},
+        error_text=metric_result.error,
+        latency_ms=latency_ms,
+    )
+
+
 def build_results(run_id: UUID, params: EvalParams, run_result: RunResult) -> list[Result]:
     """Flatten ``run_result.trace_results[*].metric_results[*]`` into a list
     of persistable :class:`Result` rows.
@@ -41,13 +85,12 @@ def build_results(run_id: UUID, params: EvalParams, run_result: RunResult) -> li
     out: list[Result] = []
     for trace_result in run_result.trace_results:
         item_id = trace_result.trace_id
-        item_name = trace_result.trace_id
         for mr in trace_result.metric_results:
             out.append(
-                Result.from_metric_result(
+                result_from_metric_result(
                     run_id=run_id,
                     eval_set_item_id=item_id,
-                    eval_set_item_name=item_name,
+                    eval_set_item_name=item_id,
                     trace_id=trace_result.trace_id,
                     evaluator_type=classify_evaluator(mr.metric_name, params),
                     metric_result=mr,
