+++ src/agentevals/storage/models.py
@@ -121,49 +121,3 @@ class Result(BaseModel):
     tokens_used: dict[str, Any] | None = None
     latency_ms: int | None = None
     created_at: datetime = Field(default_factory=_now)
-
-    @classmethod
-    def from_metric_result(
-        cls,
-        *,
-        run_id: UUID,
-        eval_set_item_id: str,
-        eval_set_item_name: str,
-        trace_id: str | None,
-        evaluator_type: Literal["builtin", "code", "remote", "openai_eval"],
-        metric_result: Any,
-    ) -> Result:
-        """Project an in-pipeline MetricResult onto the persisted shape.
-
-        ADK emits ``eval_status`` strings ``PASSED`` / ``FAILED`` /
-        ``NOT_EVALUATED``; we additionally map presence of ``error`` to
-        ``errored`` so downstream consumers don't have to special-case
-        evaluator failures.
-        """
-        if metric_result.error:
-            status = ResultStatus.ERRORED
-        else:
-            raw = (metric_result.eval_status or "NOT_EVALUATED").upper()
-            status = {
-                "PASSED": ResultStatus.PASSED,
-                "FAILED": ResultStatus.FAILED,
-            }.get(raw, ResultStatus.SKIPPED)
-
-        scores: list[float | None] = list(metric_result.per_invocation_scores or [])
-        latency_ms = int(metric_result.duration_ms) if metric_result.duration_ms is not None else None
-
-        return cls(
-            result_id=compute_result_id(run_id, eval_set_item_id, metric_result.metric_name),
-            run_id=run_id,
-            eval_set_item_id=eval_set_item_id,
-            eval_set_item_name=eval_set_item_name,
-            evaluator_name=metric_result.metric_name,
-            evaluator_type=evaluator_type,
-            status=status,
-            score=metric_result.score,
-            per_invocation_scores=scores,
-            trace_id=trace_id,
-            details=metric_result.details or {},
-            error_text=metric_result.error,
-            latency_ms=latency_ms,
-        )
